+++ Intructions.rtf
@@ -1,19 +1,51 @@
 {\rtf1\ansi\ansicpg1252\cocoartf2513
-\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica;}
-{\colortbl;\red255\green255\blue255;}
-{\*\expandedcolortbl;;}
-\margl1440\margr1440\vieww10800\viewh8400\viewkind0
+\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica-Bold;\f1\fswiss\fcharset0 Helvetica;\f2\fnil\fcharset0 Menlo-Regular;
+}
+{\colortbl;\red255\green255\blue255;\red0\green0\blue0;}
+{\*\expandedcolortbl;;\csgray\c0;}
+\margl1440\margr1440\vieww17540\viewh8400\viewkind0
 \pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\pardirnatural\partightenfactor0
 
-\f0\fs24 \cf0 Instructions on how to sort a file dcontaining a list of names:\
-\
+\f0\b\fs28 \cf0 Instructions on sorting a file
+\f1\b0 \
+
+\fs24 \
 Within the terminal, navigate to the the directory SortMeProgram containing the following files:\
-Sort Me.txt\
+Sort_Me.txt\
 Instructions\
-sortList.py\
-callSort.py\
+sortListAL.py\
+sortListRev.py\
+callSortAL.py\
+callSortRev.py\
+output1.txt\
+output2.txt\
+SortedReversed.txt\
+Sorted.txt\
+Test.sh\
 \
-Once you know you are in the right path location, type \'93python callSort.py\'94 \
+User will have two kinds of ways they can sort the contents of Sort_Me.txt\
 \
-When running the program, the user will have an option to reverse the sort. The user must press\
-1 in order for the sort to be reversed. }
+Once you know you are in the right path location: \
+\
+type \'93python callSortAL.py\'94 to sort the contents of Sort_Me.txt first by alphabetical letter and then by length of the string\
+\
+type \'93python callSortRev.py\'94 to sort the contents of Sort_Me.txt first by alphabetical letter and then by length of the string and then 
+\f0\b reverse
+\f1\b0  the order\
+\
+_________________________________________________________________________\
+\
+
+\f0\b\fs28 Instructions on testing: 
+\f1\b0\fs24 \
+\
+When you have navigated to the file location within the terminal,\
+
+\fs26 \
+Type \'93 cat Sort_Me.txt | python callSortAL.py > output1.txt \'93\
+Type  \'93 \cf2 \CocoaLigature0 cat Sort_Me.txt | python callSortRev.py > output2.txt \'93\
+\
+And to run the test, type \'93 
+\f2\fs22 ./test.sh \'94\
+\
+You should get an output saying whether each sort passed the test}
