+++ Sorted.txt
@@ -1,0 +1,100 @@
+Kha
+
+Asby
+
+Bain
+
+Dean
+
+Fife
+
+Wile
+
+Baker
+
+Ellis
+
+Evans
+
+Foley
+
+Glock
+
+Graff
+
+Heigl
+
+Lundy
+
+McVey
+
+Nylon
+
+Peery
+
+Reyes
+
+White
+
+Adkins
+
+Hickey
+
+Laymon
+
+Rogers
+
+Tanton
+
+Taylor
+
+Broome 
+
+Byfield
+
+Dulaney
+
+Hagberg
+
+Hillman
+
+McCrave
+
+Michael
+
+Padgett
+
+Routson
+
+Starkey
+
+Stegman
+
+Bostwick
+
+Cachedon
+
+Giddings
+
+Guenther
+
+Hatfield
+
+Kalivoda
+
+Kirkland
+
+Phillips
+
+Reynolds
+
+Sullivan
+
+Williams
+
+Clevenger
+
+Fitzjerrell
+
+Hendrickson
+
