--- callSort.py
@@ -1,8 +1,0 @@
-import sortList;
-
-finalsort = sortList.sort_list();
-
-for name in finalsort:
-    print(name)
-
-
