+++ callSortAL.py
@@ -1,0 +1,8 @@
+import sortListAL;
+
+finalsort = sortListAL.sort_list();
+
+for name in finalsort:
+    print(name)
+
+
