+++ callSortRev.py
@@ -1,0 +1,8 @@
+import sortListRev;
+
+finalsort = sortListRev.sort_list();
+
+for name in finalsort:
+    print(name)
+
+
